src/app/subtotal/subtotal.component.spec.ts
import { HttpClient, HttpClientModule } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { ShoppingCartService } from '../services/shopping-cart.service';

import { SubtotalComponent } from './subtotal.component';

describe('SubtotalComponent', () => {
  let component: SubtotalComponent;
  let fixture: ComponentFixture<SubtotalComponent>;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      declarations: [ SubtotalComponent ],
      imports: [ HttpClientModule,HttpClientTestingModule ],
      providers:[ShoppingCartService,HttpClient]
    })
    .compileComponents();
  });

  beforeEach(() => {
    fixture = TestBed.createComponent(SubtotalComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });
  it('should render count', () => {
    expect(component.count).toBeDefined();
  });
  it('should render subtotal', () => {
    expect(component.subTotal).toBeDefined();
  });
});
